+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Section Order:
 
 ### Changed
 
+- Use SDE for ship type and solar system in FAT model
 - Manual FAT form restructured
 - FAT link detail view is now a bit more compact, so the FAT link and the copy button don't look out of place now